fix(train): validation split read int64 tensors as raw bytes
bytes() on a numpy array returns its raw buffer, so bytes(int64_tensor.numpy())
yields eight bytes per element -- each real byte followed by seven zeros.
Splitting that on newlines compared garbage strings, and the resulting "unseen
lines" join came back 8.00x larger than the span it was filtered from
(16,683,944 bytes out of a 2,085,493-byte held-out split). Validation was
therefore scoring the model on a mostly-zero stream it had never trained on,
which is why held-out CE read 8.81 nats (12.71 BPC) -- worse than uniform --
while training CE sat at 0.31.

Cast to uint8 before taking bytes. This also retracts the causal half of DATA-1:
the corpus really is 58.7% duplicate bytes (measured on the file itself), but
the 12.3 BPC train/validation gap was this bug, not the duplication.

Note: this commit also carries the parallel session's in-flight --seed plumbing
from the shared working tree; that work is theirs, not mine.

+++ train_conscious_lm.py
@@ -878,9 +878,15 @@ def train(args: argparse.Namespace):
     # in the training split -- measured at step 20,000, that inflated the reported
     # score 2.4x: 0.594 BPC on the raw held-out span versus 1.401 BPC on lines the
     # model had never seen. Keep only the lines that are genuinely unseen.
-    train_lines = set(bytes(train_data.numpy()).split(b"\n"))
+    # bytes() reads a numpy array's RAW BUFFER, so an int64 tensor yields 8 bytes
+    # per element -- the real byte followed by seven zeros. Splitting that on b"\n"
+    # compares garbage, and the "unseen" join came back 8x larger than the span it
+    # was filtered from (16,683,944 bytes out of a 2,085,493-byte held-out split,
+    # ratio exactly 8.00), so validation was scoring the model on a mostly-zero
+    # stream it had never seen. Cast to uint8 first.
+    train_lines = set(bytes(train_data.to(torch.uint8).numpy()).split(b"\n"))
     unseen = b"\n".join(
-        ln for ln in bytes(val_raw.numpy()).split(b"\n")
+        ln for ln in bytes(val_raw.to(torch.uint8).numpy()).split(b"\n")
         if len(ln) > 8 and ln not in train_lines
     )
     if len(unseen) > (1 << 18):        # keep the fallback honest, not silent
